+++ pymc/tests/models.py
@@ -163,32 +163,6 @@ def mv_simple_discrete():
     return model.initial_point(), model, (mu, C)
 
 
-def mv_prior_simple():
-    n = 3
-    noise = 0.1
-    X = np.linspace(0, 1, n)[:, None]
-
-    K = pm.gp.cov.ExpQuad(1, 1)(X).eval()
-    L = np.linalg.cholesky(K)
-    K_noise = K + noise * np.eye(n)
-    obs = floatX_array([-0.1, 0.5, 1.1])
-
-    # Posterior mean
-    L_noise = np.linalg.cholesky(K_noise)
-    alpha = np.linalg.solve(L_noise.T, np.linalg.solve(L_noise, obs))
-    mu_post = np.dot(K.T, alpha)
-
-    # Posterior standard deviation
-    v = np.linalg.solve(L_noise, K)
-    std_post = (K - np.dot(v.T, v)).diagonal() ** 0.5
-
-    with pm.Model() as model:
-        x = pm.Flat("x", size=n)
-        x_obs = pm.MvNormal("x_obs", observed=obs, mu=x, cov=noise * np.eye(n))
-
-    return model.initial_point(), model, (K, L, mu_post, std_post, noise)
-
-
 def non_normal(n=2):
     with pm.Model() as model:
         pm.Beta("x", 3, 3, size=n, transform=None)